convert pointer to long when viewing as integer, per C standard

+++ print.c
@@ -254,10 +254,10 @@ void real_print_node(FILE *strm, NODE *nd, int depth, int width) {
 	print_char(strm,']');
     } else if (nd == UNBOUND) {
 	ndprintf(strm, "%s", theName(Name_nothing));
-    } else if ((unsigned int)nd < 200) {    /* for debugging */
+    } else if ((unsigned long)nd < 200) {    /* for debugging */
 	char num[] = "{small}    ";
 
-	sprintf(&num[7],"%d", (int)nd);
+	sprintf(&num[7],"%ld", (long)nd);
 	ndprintf(strm,num);
     } else if ((ndty = nodetype(nd)) & NT_PRIM) {
 	ndprintf(strm, "PRIM");
